Refactoring code - moved connection provider code into util package
